Move some functions from elona/init.cpp to elona/main.cpp
