+++ src/elona/Android.mk
@@ -72,6 +72,7 @@ item_material.cpp \
 itemgen.cpp \
 log.cpp \
 magic.cpp \
+main.cpp \
 main_menu.cpp \
 map.cpp \
 map_cell.cpp \
