+++ src/elona/init.hpp
@@ -5,6 +5,7 @@
 namespace elona
 {
 
+void init();
 void initialize_config(const fs::path&);
 void initialize_i18n();
 void initialize_lua();
@@ -12,6 +13,5 @@ void initialize_elona();
 void init_fovlist();
 void initialize_game();
 void initialize_debug_globals();
-int run();
 
 } // namespace elona
