+++ src/elona/main.hpp
@@ -1,0 +1,10 @@
+#pragma once
+
+
+
+namespace elona
+{
+
+int run();
+
+}
