+++ src/main.cpp
@@ -1,8 +1,8 @@
+#include "elona/main.hpp"
 #include <iostream>
 #include <stdexcept>
 #include <SDL.h>
 #include "elona/defines.hpp"
-#include "elona/init.hpp"
 #include "elona/log.hpp"
 #include "version.hpp"
